+++  course-work-ps-31-2013 --username dertkaes@gmail.com/OpenGLTest/OpenGLTest/view/Table.cpp
@@ -19,7 +19,7 @@ void CTable::Draw() const
 	texManager->SetTexture("");
 }
 
-bool CTable::isCoordsOnTable(int worldX, int worldY) const
+bool CTable::isCoordsOnTable(double worldX, double worldY) const
 {
 	return (worldX < GetWidth() / 2 && worldX > -GetWidth() / 2 
 	   && worldY < GetHeight() / 2 && worldY > -GetHeight() / 2);
